New/RemovedPCREModifiers: decouple the sniffs, introduce `PCRERegexTrait`
This moves the logic which is needed by both the `NewPCREModifiers` as well as the `RemovedPCREModifiers` to a dedicated trait.

This allows us to decouple the `NewPCREModifiers` sniff from the `RemovedPCREModifiersSniff` and let it extend the `AbstractFunctionCallParameterSniff` class instead.

Sniffs which extend other sniffs can run into problems with the autoloading used in PHPCS as per the reports in issue 638 and 793.
This is part of an effort to fix those problems.

Having this trait in place will also make it easier to potentially, eventually address issue 965.

This commit does not change the actual functionality of the sniffs.
It only moves the relevant code to the trait and makes the trait methods stand-alone methods which return a value instead of passing off directly to the next function.

The only difference is that now the last token in the regex pattern will be used to report the error on in all cases.
Previously, this was only the case for array parameters, not for string parameters.

Includes adjusted line numbers in the test files for this.