PCRERegexTrait: use PHPCSUtils `getDoubleArrowPtr()` method
This will be more reliable than just searching for a double arrow as, while unlikely in this particular case, just searching for a double arrow could incorrectly identify an arrow of a nested array, the arrow in a `yield` or for an arrow function as the array double arrow.

The `Arrays::getDoubleArrowPtr()` method contains protection against all these type of issues.