PCRERegexTrait: use PHPCSUtils `Collections::$arrayTokensBC`
To check whether the paremeter is an array, we can use the `Collections::$arrayTokensBC` array, which also includes the `T_OPEN_SQUARE_BRACKET` which in some cases is in actual fact a short array when not correctly tokenized.

No need to follow this by a call to `Arrays::isShortArray()` as the regex parameter can't really be a short list as that doesn't make any sense at all.